Refactor HomePageCarousel out of Home
