+++ frontends/mit-open/src/pages/Home/Home.tsx
@@ -1,86 +1,12 @@
 import React, { useCallback, useState } from "react"
-import {
-  Button,
-  ChipLink,
-  Container,
-  Grid,
-  useMuiBreakpointAtLeast,
-  SearchInput,
-} from "ol-design"
-import type { SearchInputProps } from "ol-design"
-import { GridContainer } from "../components/layout"
-import { TitledCarousel } from "ol-util"
-import ArrowBack from "@mui/icons-material/ArrowBack"
-import ArrowForward from "@mui/icons-material/ArrowForward"
-import LearningResourceCard from "../components/LearningResourceCard"
-import type { PaginatedLearningResourceList } from "api"
-import { useLearningResourcesList } from "api/hooks/learningResources"
-import { UseQueryResult } from "@tanstack/react-query"
 import styled from "@emotion/styled"
-import { Theme } from "../entry/theme"
+import { ChipLink, Container, Grid, SearchInput } from "ol-design"
+import type { SearchInputProps } from "ol-design"
 import { mediaQueries } from "ol-util"
-
-interface HomePageCarouselProps {
-  query: UseQueryResult<PaginatedLearningResourceList>
-  showNavigationButtons?: boolean
-  title: React.ReactNode
-}
-
-const CarouselButton = styled(Button)`
-  padding: 10px;
-  padding-left: 15px;
-  padding-right: 15px;
-  margin-right: 0.5em;
-  margin-left: 0.5em;
-`
-
-const HomePageCarousel: React.FC<HomePageCarouselProps> = ({
-  query,
-  showNavigationButtons = true,
-  title,
-}) => {
-  const aboveSm = useMuiBreakpointAtLeast("sm")
-  const aboveLg = useMuiBreakpointAtLeast("lg")
-  const pageSize = aboveLg ? 4 : aboveSm ? 2 : 1
-
-  return (
-    <TitledCarousel
-      title={title}
-      as="section"
-      pageSize={pageSize}
-      carouselClassName="ic-carousel"
-      cellSpacing={0} // we'll handle it with css
-      previous={
-        <CarouselButton
-          variant="outlined"
-          color="secondary"
-          startIcon={<ArrowBack fontSize="inherit" />}
-        >
-          Previous
-        </CarouselButton>
-      }
-      next={
-        <CarouselButton
-          variant="outlined"
-          color="secondary"
-          endIcon={<ArrowForward fontSize="inherit" />}
-        >
-          Next
-        </CarouselButton>
-      }
-      showNavigationButtons={showNavigationButtons}
-    >
-      {query.data?.results?.map((resource) => (
-        <LearningResourceCard
-          key={resource.id}
-          className="ic-resource-card ic-carousel-card"
-          resource={resource}
-          variant="column"
-        />
-      ))}
-    </TitledCarousel>
-  )
-}
+import { GridContainer } from "../../components/layout"
+import { useLearningResourcesList } from "api/hooks/learningResources"
+import { Theme } from "../../entry/theme"
+import HomePageCarousel from "./HomePageCarousel"
 
 const EXPLORE_BUTTONS = [
   {
